backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
sqlalchemy
databases
alembic
python-multipart
passlib[bcrypt]
PyJWT
email-validator
python-dotenv
